requirements.txt
langchain==0.3.7
langchain-groq==0.2.1
python-dotenv==1.0.1
langgraph>=0.2.0
langchain-community
